Add initial upstream code for srfi-207: String-notated bytevectors
https://srfi.schemers.org/srfi-207/srfi-207.html

Add the key files from the upstream implementation, excepting the
Guile-specific top-level module to integrate it all. Omit
parse.scm because we're going to provide an alternative.

These files are the unmodified upstream code from the upstream
repository:

  https://github.com/scheme-requests-for-implementation/srfi-207.git

as of this commit (errata-2-2-gd646cfa):

  commit d646cfa6171912398e9325862ba1e72d75f301f9
  Merge: 692530f 21a5c30
  Author: Arthur A. Gleckler <srfi@speechcode.com>
  Date:   Fri Feb 21 19:28:07 2025 -0800

      Merge pull request #16 from Zipheir/master

      Fix incorrect example.

For now add the html documentation as-is; we'll convert it to texinfo
later.

* doc/ref/srfi-207.html: Add new file.
* module/srfi/srfi-207/upstream/base64.scm: Add new file.
* module/srfi/srfi-207/upstream/bytestrings-impl.scm: Add new file.
* test-suite/tests/srfi-207.test: Add new file.