make-struct/no-tail scm_c_make_structv: ensure layout is symbol
Previously we just tested for a layout via scm_is_true (SCM_LAYOUT ...),
but layouts are symbols, and that test allowed us to proceed to the code
in scm_is_valid_vtable_layout that called scm_i_symbol_chars on the
"layout", segfaulting if the layout was say an integer, as it would be
for something like this:

  (make-struct/no-tail (make-vtable standard-vtable-fields #f) 0)

Since a layout must be a symbol, check via scm_is_symbol instead.

* libguile/struct.c (scm_c_make_structv): test for layout via
scm_is_symbol.

Thanks to Ido Yariv for reporting the problem and Dale P. Smith for help
determining the cause.

Closes: 78789