tmpfile: preserve O_CLOEXEC so children won't prevent removal
Guile duplicates the underlying file descriptor stored in a FILE pointer
returned by tmpfile(3).  This is done to manage the file descriptor with
Guile's ports abstraction instead of using the C FILE pointer
abstraction.

However, the dup(2) system call does not copy the file descriptor
flags (O_CLOEXEC), which is important for respecting the semantic
expected from tmpfile(3) in that the temporary file will not leak to
child processes.

Fix this by using the fcntl(2) system call using the `F_DUPFD_CLOEXEC'
command.

* libguile/posix.c (scm_tmpfile): Use fcntl(2) instead of dup(2) when
not on MINGW32 systems.

[rlb@defaultvalue.org: adjust commit summary and tmpfile comment; add NEWS]